style: Protocol과 Error를 별도의 폴더로 분리
